Split handle types into their own header
Move LiteralHandle/TemporaryHandle/LocalHandle/GlobalHandle, the Handle
variant, and get_raw_handle out of instructions.hpp into a new handle.hpp.
