+++ bust/codegen/instructions.hpp
@@ -11,67 +11,13 @@
 #pragma once
 //****************************************************************************
 
+#include "codegen/handle.hpp"
 #include "codegen/types.hpp"
-#include <string>
 #include <variant>
 
 //****************************************************************************
 namespace bust::codegen {
 //****************************************************************************
-
-struct LiteralHandle {
-  std::string m_handle;
-};
-
-struct TemporaryHandle {
-  static size_t m_unique_id;
-  size_t m_id = m_unique_id++;
-};
-
-struct LocalHandle {
-  std::string m_handle;
-};
-
-struct GlobalHandle {
-  std::string m_handle;
-};
-
-using Handle =
-    std::variant<LiteralHandle, TemporaryHandle, LocalHandle, GlobalHandle>;
-
-// inline std::string handle_to_string(const Handle &handle) {
-//   return std::visit(
-//       [](const auto &handle) -> std::string {
-//         using T = std::decay_t<decltype(handle)>;
-//         if constexpr (std::is_same_v<T, LiteralHandle>) {
-//           return handle.m_handle;
-//         } else if constexpr (std::is_same_v<T, TemporaryHandle>) {
-//           return "%" + std::to_string(handle.m_id);
-//         } else if constexpr (std::is_same_v<T, LocalHandle>) {
-//           return "%" + handle.m_handle;
-//         } else if constexpr (std::is_same_v<T, GlobalHandle>) {
-//           return "@" + handle.m_handle;
-//         }
-//       },
-//       handle);
-// }
-
-inline std::string get_raw_handle(const Handle &handle) {
-  return std::visit(
-      [](const auto &handle) -> std::string {
-        using T = std::decay_t<decltype(handle)>;
-        if constexpr (std::is_same_v<T, TemporaryHandle>) {
-          return std::to_string(handle.m_id);
-        } else {
-          return handle.m_handle;
-        }
-      },
-      handle);
-}
-
-// inline std::ostream &operator<<(std::ostream &out, const Handle &handle) {
-//   return out << handle_to_string(handle);
-// }
 
 struct BinaryInstruction {
   Handle m_result;
